# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/specialty/create": {
      "post": {
        "operationId": "SpecialtyController_createSpecialty",
        "summary": "Create a new specialty",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpecialtyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Specialty created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Specialty created successfully"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "name": ["name must be a string"]
                  }
                }
              }
            }
          }
        },
        "tags": ["specialty"]
      }
    },
    "/specialty/get/{id}": {
      "get": {
        "operationId": "SpecialtyController_getSpecialty",
        "summary": "Get specialty by ID or all specialties",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "description": "Specialty ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Specialty(s) retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "123",
                      "name": "Specialty 1"
                    },
                    {
                      "id": "124",
                      "name": "Specialty 2"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Invalid specialty ID"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Specialty not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Specialty not found"
                  }
                }
              }
            }
          }
        },
        "tags": ["specialty"]
      }
    },
    "/specialty/get": {
      "get": {
        "operationId": "SpecialtyController_getSpecialty",
        "summary": "Get specialty by ID or all specialties",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "description": "Specialty ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Specialty(s) retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "123",
                      "name": "Specialty 1"
                    },
                    {
                      "id": "124",
                      "name": "Specialty 2"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Invalid specialty ID"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Specialty not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Specialty not found"
                  }
                }
              }
            }
          }
        },
        "tags": ["specialty"]
      }
    },
...
    "/specialty/update/{id}": {
      "patch": {
        "operationId": "SpecialtyController_updateSpecialty",
        "summary": "Update specialty by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Specialty ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpecialtyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Specialty updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Specialty updated successfully"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Specialty not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Specialty not found"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "name": ["name must be a string"]
                  }
                }
              }
            }
          }
        },
        "tags": ["specialty"]
      }
    },
    "/specialty/delete/{id}": {
      "delete": {
        "operationId": "SpecialtyController_deleteSpecialty",
        "summary": "Delete specialty by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Specialty ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Specialty deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Specialty deleted successfully"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Specialty not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Specialty not found"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": ["Invalid specialty ID"]
                  }
                }
              }
            }
          }
        },
        "tags": ["specialty"]
      }
    },
...
  },
  "info": {
    "title": "Career API",
    "description": "The Career API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSpecialtyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Computer Science",
            "description": "The name of the specialty"
          },
          "EIOHPE": {
            "type": "string",
            "example": "EIOHPE123",
            "description": "Educational institutions of higher professional education"
          },
          "class": {
            "type": "number",
            "example": 3,
            "description": "The class level of the specialty"
          },
          "specializationGroup": {
            "type": "number",
            "example": 2,
            "description": "The specialization group of the specialty"
          },
          "clusterName": {
            "type": "string",
            "example": "Engineering",
            "description": "The cluster name of the specialty"
          },
          "clusterTag": {
            "type": "string",
            "example": "ENG",
            "description": "The cluster tag of the specialty"
          },
          "specialtyDescription": {
            "type": "string",
            "example": "A specialization in computer science",
            "description": "The description of the specialty"
          },
          "specialtyCode": {
            "type": "number",
            "example": 101,
            "description": "The specialty code"
          },
          "specialtyName": {
            "type": "string",
            "example": "Bachelor of Computer Science",
            "description": "The name of the specialty"
          },
          "formOfEducation": {
            "type": "string",
            "example": "Full-time",
            "description": "The form of education"
          },
          "typeOfStudy": {
            "type": "string",
            "example": "On-campus",
            "description": "The type of study"
          },
          "languageOfStudy": {
            "type": "string",
            "example": "English",
            "description": "The language of study"
          },
          "universityName": {
            "type": "string",
            "example": "XYZ University",
            "description": "The university name"
          },
          "monthlyIncome": {
            "type": "number",
            "example": 1000,
            "description": "The monthly income from the specialty"
          },
          "skillsLevel": {
            "type": "number",
            "example": 5,
            "description": "The skills level required for the specialty"
          },
          "futureGrowth": {
            "type": "string",
            "example": "High",
            "description": "Future growth opportunities for the specialty"
          },
          "overview": {
            "type": "string",
            "example": "An overview of the specialty",
            "description": "The overview of the specialty"
          },
          "careerOpportunities": {
            "example": ["Software Development", "Data Analysis"],
            "description": "Career opportunities available",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "EIOHPE",
          "class",
          "specializationGroup",
          "clusterName",
          "clusterTag",
          "specialtyDescription",
          "specialtyCode",
          "specialtyName",
          "formOfEducation",
          "typeOfStudy",
          "languageOfStudy",
          "universityName",
          "monthlyIncome",
          "skillsLevel",
          "futureGrowth",
          "overview",
          "careerOpportunities"
        ]
      },
      "UpdateSpecialtyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Computer Science",
            "description": "The name of the specialty"
          },
          "EIOHPE": {
            "type": "string",
            "example": "EIOHPE123",
            "description": "Educational institutions of higher professional education"
          },
          "class": {
            "type": "number",
            "example": 3,
            "description": "The class level of the specialty"
          },
          "specializationGroup": {
            "type": "number",
            "example": 2,
            "description": "The specialization group of the specialty"
          },
          "clusterName": {
            "type": "string",
            "example": "Engineering",
            "description": "The cluster name of the specialty"
          },
          "clusterTag": {
            "type": "string",
            "example": "ENG",
            "description": "The cluster tag of the specialty"
          },
          "specialtyDescription": {
            "type": "string",
            "example": "A specialization in computer science",
            "description": "The description of the specialty"
          },
          "specialtyCode": {
            "type": "number",
            "example": 101,
            "description": "The specialty code"
          },
          "specialtyName": {
            "type": "string",
            "example": "Bachelor of Computer Science",
            "description": "The name of the specialty"
          },
          "formOfEducation": {
            "type": "string",
            "example": "Full-time",
            "description": "The form of education"
          },
          "typeOfStudy": {
            "type": "string",
            "example": "On-campus",
            "description": "The type of study"
          },
          "languageOfStudy": {
            "type": "string",
            "example": "English",
            "description": "The language of study"
          },
          "universityName": {
            "type": "string",
            "example": "XYZ University",
            "description": "The university name"
          },
          "monthlyIncome": {
            "type": "number",
            "example": 1000,
            "description": "The monthly income from the specialty"
          },
          "skillsLevel": {
            "type": "number",
            "example": 5,
            "description": "The skills level required for the specialty"
          },
          "futureGrowth": {
            "type": "string",
            "example": "High",
            "description": "Future growth opportunities for the specialty"
          },
          "overview": {
            "type": "string",
            "example": "An overview of the specialty",
            "description": "The overview of the specialty"
          },
          "careerOpportunities": {
            "example": ["Software Development", "Data Analysis"],
            "description": "Career opportunities available",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}